Migrate test/* test files to Vitest
